Adding e calculations necessary for the booking
Defining help function:

Calculate the number of available seats for a certain flight: Function call:
calculateFreeSeats(flightnumber); where the output is the number of free
(unpaid) seats (i.e. an integer) on that certain flight.

Calculate the price of the next seat on a flight: Function call:
calculatePrice(flightnumber); where the output is the price (i.e. a double) of
the next seat calculated as shown in 1e